ACOG Districts Data

This dataset contains information about American College of Obstetricians and Gynecologists (ACOG) districts, including their two-letter state abbreviations and full state names.

Usage

ACOG_Districts

Format

A data frame with the following columns:

ACOG_District

Two-letter state abbreviations representing ACOG districts.

name

Full names of the states corresponding to the ACOG districts.

Source

Data was obtained from the official ACOG website: https://www.acog.org/community/districts-and-sections

Value

A tibble where each row represents an ACOG district with its corresponding two-letter abbreviation and full state name.

See also

Examples

# Load the ACOG Districts Data
data(ACOG_Districts)

# Inspect the dataset
print(ACOG_Districts)
#> # A tibble: 52 × 4
#>    State                ACOG_District Subregion     State_Abbreviations
#>    <chr>                <chr>         <chr>         <chr>              
#>  1 Alabama              District VII  District VII  AL                 
#>  2 Alaska               District VIII District VIII AK                 
#>  3 Arizona              District VIII District VIII AZ                 
#>  4 Arkansas             District VII  District VII  AR                 
#>  5 California           District IX   District IX   CA                 
#>  6 Colorado             District VIII District VIII CO                 
#>  7 Connecticut          District I    District I    CT                 
#>  8 Delaware             District IV   District IV   DE                 
#>  9 District of Columbia District IV   District IV   DC                 
#> 10 Florida              District XII  District XII  FL                 
#> # ℹ 42 more rows

# Get a summary of the dataset
summary(ACOG_Districts)
#>     State           ACOG_District       Subregion         State_Abbreviations
#>  Length:52          Length:52          Length:52          Length:52          
#>  Class :character   Class :character   Class :character   Class :character   
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character   

# Perform data analysis and exploration