Create a Formula for Poisson Model

This function creates a formula for a Poisson model based on the provided data, response variable, and optional random effect.

Usage

create_formula(data, response_var, random_effect = NULL)

Arguments

data

A dataframe containing the predictor and response variables.

response_var

The name of the response variable in the dataframe.

random_effect

Optional. The name of the random effect variable for the formula.

Value

A formula object suitable for modeling in R.

Examples

# Example usage:
response_variable <- "days"
random_effect_term <- "name"  # Change this to the desired random effect variable
df3_filtered <- data.frame(days = c(5, 10, 15), age = c(30, 40, 50), name = c("A", "B", "C"))
formula <- create_formula(df3_filtered, response_variable, random_effect_term)
#> Creating formula with response variable: days
#> Predictor variables identified: age, name
#> Predictor variables after formatting: `age`, `name`
#> Initial formula string: days ~ `age` + `name`
#> Formula string with random effect: days ~ `age` + `name` + (1 | name )
#> Final formula object created: days ~ age + name + (1 | name)
formula
#> days ~ age + name + (1 | name)
#> <environment: 0x7fd34d53d9b0>