Main function to create a map of U.S. States by Region, Division, or Custom Districts

This function generates a map of U.S. states, coloring them by regions, divisions, ACOG districts, or ENT Board of Governors Regions. Users can customize which grouping to use (ACOG Districts, ENT_Board_of_Governors_Regions, or US Census Subdivisions).

Usage

create_region_map(
  remove_ak_hi = TRUE,
  districts_per_group = "acog_districts",
  save_path = NULL,
  alpha_level = 0.4,
  title = "U.S. States by Region/Division or Custom Districts",
  subtitle = NULL
)

Arguments

remove_ak_hi

Logical, whether to exclude Alaska and Hawaii from the map. Default is TRUE.

districts_per_group

Character, which grouping to use for the regions. Options are: "acog_districts", "ENT_Board_of_Governors_Regions", or "US_Census_Subdivisions".

save_path

Character, an optional file path to save the map image. If NULL, the map will not be saved. Default is NULL.

alpha_level

Numeric, the transparency level for the map's fill color, with 0 being fully transparent and 1 fully opaque. Default is 0.2.

title

Character, the title for the map. Default is "U.S. States by Region/Division or Custom Districts".

subtitle

Character, an optional subtitle for the map.

Value

A ggplot map object representing the U.S. states colored by region/division, ACOG districts, or ENT_Board_of_Governors_Regions.

Examples

# Example 1: Create a map of U.S. states by region/division, excluding Alaska and Hawaii
create_region_map(remove_ak_hi = TRUE, districts_per_group = "US_Census_Subdivisions", alpha_level = 0.2)
#> Logging inputs...
#> Error in validate_inputs(remove_ak_hi, districts_per_group, alpha_level,     save_path): Error: 'significant_predictors' must be a data frame.

# Example 2: Create a map of U.S. states by ACOG districts, including Alaska and Hawaii
create_region_map(remove_ak_hi = FALSE, districts_per_group = "acog_districts", alpha_level = 0.5)
#> Logging inputs...
#> Error in validate_inputs(remove_ak_hi, districts_per_group, alpha_level,     save_path): Error: 'significant_predictors' must be a data frame.

# Example 3: Create a map of U.S. states by ENT Board of Governors Regions and save it to a file
create_region_map(remove_ak_hi = TRUE, districts_per_group = "ENT_Board_of_Governors_Regions", save_path = "ent_bog_regions_map.png", alpha_level = 0.3)
#> Logging inputs...
#> Error in validate_inputs(remove_ak_hi, districts_per_group, alpha_level,     save_path): Error: 'significant_predictors' must be a data frame.