Get Hospital Referral Region Shapefile

This function loads the hospital referral region shapefile and optionally removes Hawaii and Alaska. The shapefile (~8 MB) is downloaded from the Dartmouth Atlas website on first use and cached in the user's R cache directory for subsequent calls.

Usage

hrr(remove_HI_AK = TRUE)

Arguments

remove_HI_AK

Logical, should Hawaii and Alaska be removed? Default is TRUE.

Value

An sf object containing the hospital referral region data.

Examples

hrr()
#> Loading necessary packages...
#> Getting the hospital referral region shapefile...
#> Downloading HRR boundary shapefile (~8 MB). This is a one-time operation.
#> Hospital Referral Region shapefile loaded.
#> This function creates an sf file of hospital referral regions.
#> For more information: https://data.dartmouthatlas.org/supplemental/...
#> Simple feature collection with 304 features and 7 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -124.7319 ymin: 24.84407 xmax: -66.95047 ymax: 49.38308
#> Geodetic CRS:  WGS 84
#> # A tibble: 304 × 8
#>    hrrnum hrrcity        stf3bpop color color8a color8b HRR_lbl   
#>  *  <int> <chr>             <int> <int>   <int>   <int> <chr>     
#>  1      1 AL- BIRMINGHAM        0     1       7       7 Birmingham
#>  2      2 AL- DOTHAN            0    15       2       3 Dothan    
#>  3      5 AL- HUNTSVILLE        0    10       5       5 Huntsville
#>  4      6 AL- MOBILE            0     1       4       2 Mobile    
#>  5      7 AL- MONTGOMERY        0     9       5       5 Montgomery
#>  6      9 AL- TUSCALOOSA        0    10       8       3 Tuscaloosa
#>  7     11 AZ- MESA              0    12       5       4 Mesa      
#>  8     12 AZ- PHOENIX           0     8       3       6 Phoenix   
#>  9     14 AZ- SUN CITY          0     1       7       3 Sun City  
#> 10     15 AZ- TUCSON            0     5       6       1 Tucson    
#> # ℹ 294 more rows
#> # ℹ 1 more variable: geometry <MULTIPOLYGON [°]>