Perform Logistic Regression on Multiple Predictors with Logging

This function performs logistic regression for multiple predictor variables against a specified target variable. It logs the process, including inputs, data transformations, and results. The function filters predictors based on a given significance level.

logistic_regression(df, target_variable, predictor_vars, significance_level)

Arguments

df

A data frame containing the target and predictor variables.

target_variable

A string representing the name of the target variable.

predictor_vars

A vector of strings representing the names of predictor variables.

significance_level

A numeric value specifying the significance level for filtering predictors.

Value

A data frame containing significant predictors with their p-values and formatted p-values.

Examples

# Assuming df is your data frame
target_variable <- "cleaned_does_the_physician_accept_medicaid_numeric"
predictor_vars <- setdiff(
  names(df),
  c(
    target_variable,
    "does_the_physician_accept_medicaid",
    "cleaned_does_the_physician_accept_medicaid"
  )
)
significance_logistic_regression <- 0.2

significant_vars <- logistic_regression(
  df,
  target_variable,
  predictor_vars,
  significance_level = significance_logistic_regression
)
#> Starting logistic_regression...
#> Target Variable: cleaned_does_the_physician_accept_medicaid_numeric 
#> Predictor Variables:  
#> Significance Level: 0.2 
#> Significant Predictors:
#> [1] Variable          P_Value           Formatted_P_Value
#> <0 rows> (or 0-length row.names)
print(significant_vars)
#> [1] Variable          P_Value           Formatted_P_Value
#> <0 rows> (or 0-length row.names)