Map Physicians by State or Subdivision

This function generates a choropleth map based on the physician counts and saves the map as a PNG file.

map_physicians_by_group(
  state_counts,
  output_file_prefix = "choropleth_map",
  group_by = "state"
)

Arguments

state_counts

A tibble containing the counts of physicians per state or subdivision.

output_file_prefix

A prefix for the output PNG file of the choropleth map (default is "choropleth_map").

group_by

A string indicating whether the counts are grouped by "state" or "subdivision" (default is "state").

Value

A ggplot2 object representing the choropleth map.

Examples

# Example 1: Map physicians by state
state_counts <- count_physicians_by_group(taxonomy_and_aaos_data)
#> Error: Failed to evaluate glue component {nrow(data)}
#> Caused by error:
#> ! object 'taxonomy_and_aaos_data' not found
map_physicians_by_group(state_counts)
#> INFO [2024-11-18 00:43:03] Mapping physicians grouped by state
#> Error in merge_state_counts_with_map(state_counts): could not find function "merge_state_counts_with_map"

# Example 2: Map physicians by U.S. Census Bureau subdivision
subdivision_counts <- count_physicians_by_group(taxonomy_and_aaos_data, group_by = "subdivision")
#> Error: Failed to evaluate glue component {nrow(data)}
#> Caused by error:
#> ! object 'taxonomy_and_aaos_data' not found
map_physicians_by_group(subdivision_counts, group_by = "subdivision")
#> INFO [2024-11-18 00:43:03] Mapping physicians grouped by subdivision
#> Error in merge_subdivision_counts_with_map(state_counts): could not find function "merge_subdivision_counts_with_map"