Create a table in DuckDB from a CSV file

This function creates a table in DuckDB by reading data from a CSV file.

Usage

nppes_create_table_from_csv(con, file_path, table_name = "npi_2024")

Arguments

con

A DuckDB connection object.

file_path

A character string specifying the path to the CSV file.

table_name

A character string specifying the name of the table to be created. Default is "npi_2024".

Value

None

Details

After creating the table, the next step is to query sample data using nppes_query_sample_data().

Examples

con <- nppes_connect_to_duckdb("/path/to/your/duckdb_file.duckdb")
#> Connecting to DuckDB at /path/to/your/duckdb_file.duckdb ...
#> Warning: cannot open file '/path/to/your/duckdb_file.duckdb': No such file or directory
#> Error in file(con, "w"): cannot open the connection
nppes_create_table_from_csv(con, "/path/to/your/csv_file.csv", "npi_2024")
#> Error in nppes_create_table_from_csv(con, "/path/to/your/csv_file.csv",     "npi_2024"): could not find function "nppes_create_table_from_csv"
# Next step: nppes_query_sample_data()

con <- nppes_connect_to_duckdb("/Volumes/Video Projects Muffly 1/nppes_historical_downloads/my_duckdb.duckdb")
#> Connecting to DuckDB at /Volumes/Video Projects Muffly 1/nppes_historical_downloads/my_duckdb.duckdb ...
#> Warning: cannot open file '/Volumes/Video Projects Muffly 1/nppes_historical_downloads/my_duckdb.duckdb': No such file or directory
#> Error in file(con, "w"): cannot open the connection
nppes_create_table_from_csv(con, "/Volumes/Video Projects Muffly 1/nppes_historical_downloads/NPPES_Data.csv", "npi_2024")
#> Error in nppes_create_table_from_csv(con, "/Volumes/Video Projects Muffly 1/nppes_historical_downloads/NPPES_Data.csv",     "npi_2024"): could not find function "nppes_create_table_from_csv"
# Next step: nppes_query_sample_data()