Phase 0 Sample Creation by Region for Ortho Spine

This function processes taxonomy and AAOS data, merging it with Census Bureau data to create a sampled dataset grouped by region for ortho spine. Outputs are saved as a CSV.

phase0_create_sampled_by_region(
  taxonomy_and_aaos_tbl,
  census_bureau_tbl,
  output_csv_path = "ortho_spine/phase_0/sampled_by_region_ortho_spine.csv"
)

Arguments

taxonomy_and_aaos_tbl

A tibble containing taxonomy and AAOS data.

census_bureau_tbl

A tibble containing Census Bureau data with regional information.

output_csv_path

The file path to save the sampled dataset. Defaults to "ortho_spine/phase_0/sampled_by_region_ortho_spine.csv".

Value

A tibble of sampled ortho spine records by region.

Examples

# Example 1: Basic usage with default output path
sampled_data <- phase0_create_sampled_by_region(taxonomy_and_aaos_tbl, census_bureau_tbl)
#> INFO [2024-11-18 00:43:04] Starting phase0_create_sampled_by_region.
#> Error in (function (e) {    args <- paste0(capture.output(str(...)), collapse = "\n")    stop(paste0("`glue` failed in `formatter_glue` on:\n\n",         args, "\n\nRaw error message:\n\n", conditionMessage(e),         "\n\nPlease consider using another `log_formatter` or ",         "`skip_formatter` on strings with curly braces."))})(structure(list(message = "Failed to evaluate glue component {nrow(taxonomy_and_aaos_tbl)}",     trace = structure(list(call = list(pkgdown::build_site_github_pages(new_process = FALSE,         install = FALSE), build_site(pkg, preview = FALSE, install = install,         new_process = new_process, ...), build_site_local(pkg = pkg,         examples = examples, run_dont_run = run_dont_run, seed = seed,         lazy = lazy, override = override, preview = preview,         devel = devel), build_reference(pkg, lazy = lazy, examples = examples,         run_dont_run = run_dont_run, seed = seed, override = override,         preview = FALSE, devel = devel), unwrap_purrr_error(purrr::map(topics,         build_reference_topic, pkg = pkg, lazy = lazy, examples_env = examples_env,         run_dont_run = run_dont_run)), withCallingHandlers(code,         purrr_error_indexed = function(err) {            cnd_signal(err$parent)        }), purrr::map(topics, build_reference_topic, pkg = pkg,         lazy = lazy, examples_env = examples_env, run_dont_run = run_dont_run),         map_("list", .x, .f, ..., .progress = .progress), with_indexed_errors(i = i,             names = names, error_call = .purrr_error_call, call_with_cleanup(map_impl,                 environment(), .type, .progress, n, names, i)),         withCallingHandlers(expr, error = function(cnd) {            if (i == 0L) {            }            else {                message <- c(i = "In index: {i}.")                if (!is.null(names) && !is.na(names[[i]]) &&                   names[[i]] != "") {                  name <- names[[i]]                  message <- c(message, i = "With name: {name}.")                }                else {                  name <- NULL                }                cli::cli_abort(message, location = i, name = name,                   parent = cnd, call = error_call, class = "purrr_error_indexed")            }        }), call_with_cleanup(map_impl, environment(), .type,             .progress, n, names, i), .f(.x[[i]], ...), withCallingHandlers(data_reference_topic(topic,             pkg, examples_env = examples_env, run_dont_run = run_dont_run),             error = function(err) {                cli::cli_abort("Failed to parse Rd in {.file {topic$file_in}}",                   parent = err, call = quote(build_reference()))            }), data_reference_topic(topic, pkg, examples_env = examples_env,             run_dont_run = run_dont_run), run_examples(tags$tag_examples[[1]],             env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),             topic = tools::file_path_sans_ext(topic$file_in),             run_dont_run = run_dont_run), highlight_examples(code,             topic, env = env), downlit::evaluate_and_highlight(code,             fig_save = fig_save_topic, env = eval_env, output_handler = handler),         evaluate::evaluate(code, child_env(env), new_device = TRUE,             output_handler = output_handler), withRestarts(with_handlers({            for (expr in tle$exprs) {                ev <- withVisible(eval(expr, envir))                watcher$capture_plot_and_output()                watcher$print_value(ev$value, ev$visible, envir)            }            TRUE        }, handlers), eval_continue = function() TRUE, eval_stop = function() FALSE,             eval_error = function(cnd) {                signalCondition(cnd)                stop(cnd)            }), withRestartList(expr, restarts), withOneRestart(withRestartList(expr,             restarts[-nr]), restarts[[nr]]), doWithOneRestart(return(expr),             restart), withRestartList(expr, restarts[-nr]), withOneRestart(withRestartList(expr,             restarts[-nr]), restarts[[nr]]), doWithOneRestart(return(expr),             restart), withRestartList(expr, restarts[-nr]), withOneRestart(expr,             restarts[[1L]]), doWithOneRestart(return(expr), restart),         with_handlers({            for (expr in tle$exprs) {                ev <- withVisible(eval(expr, envir))                watcher$capture_plot_and_output()                watcher$print_value(ev$value, ev$visible, envir)            }            TRUE        }, handlers), eval(call), eval(call), withCallingHandlers(code,             message = `<fn>`, warning = `<fn>`, error = `<fn>`),         withVisible(eval(expr, envir)), eval(expr, envir), eval(expr,             envir), phase0_create_sampled_by_region(taxonomy_and_aaos_tbl,             census_bureau_tbl), logger::log_info("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns."),         log_level(INFO, ..., namespace = namespace, .logcall = .logcall,             .topcall = .topcall, .topenv = .topenv), lapply(definitions,             function(definition) {                if (level > definition$threshold) {                  return(NULL)                }                log_fun <- do.call(logger, definition)                structure(do.call(log_fun, log_arg), class = "logger")            }), FUN(X[[i]], ...), structure(do.call(log_fun,             log_arg), class = "logger"), do.call(log_fun, log_arg),         `<fn>`("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns.",             level = `<loglevel>`, .logcall = logger::log_info("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns."),             .topcall = phase0_create_sampled_by_region(taxonomy_and_aaos_tbl,                 census_bureau_tbl), .topenv = `<env>`, namespace = "tyler"),         do.call(formatter, c(res$params, list(.logcall = substitute(.logcall),             .topcall = substitute(.topcall), .topenv = .topenv))),         `<fn>`("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns.",             .logcall = logger::log_info("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns."),             .topcall = phase0_create_sampled_by_region(taxonomy_and_aaos_tbl,                 census_bureau_tbl), .topenv = `<env>`), withCallingHandlers(glue::glue(...,             .envir = .topenv), error = function(e) {            args <- paste0(capture.output(str(...)), collapse = "\n")            stop(paste0("`glue` failed in `formatter_glue` on:\n\n",                 args, "\n\nRaw error message:\n\n", conditionMessage(e),                 "\n\nPlease consider using another `log_formatter` or ",                 "`skip_formatter` on strings with curly braces."))        }), glue::glue(..., .envir = .topenv), glue_data(.x = NULL,             ..., .sep = .sep, .envir = .envir, .open = .open,             .close = .close, .na = .na, .null = .null, .comment = .comment,             .literal = .literal, .transformer = .transformer,             .trim = .trim), `<fn>`("nrow(taxonomy_and_aaos_tbl)"),         .transformer(expr, env) %||% .null, .transformer(expr,             env), with_glue_error(eval(expr, envir), paste0("Failed to evaluate glue component {",             text, "}")), withCallingHandlers(expr, error = function(cnd) {            rlang::abort(message, parent = cnd, call = NULL)        }), eval(expr, envir), eval(expr, envir), nrow(taxonomy_and_aaos_tbl),         .handleSimpleError(`<fn>`, "object 'taxonomy_and_aaos_tbl' not found",             base::quote(eval(expr, envir))), h(simpleError(msg,             call)), rlang::abort(message, parent = cnd, call = NULL)),         parent = c(0L, 1L, 2L, 3L, 4L, 5L, 4L, 7L, 8L, 9L, 8L,         8L, 12L, 12L, 14L, 15L, 16L, 17L, 18L, 19L, 20L, 21L,         20L, 23L, 24L, 23L, 26L, 27L, 18L, 29L, 30L, 29L, 18L,         18L, 34L, 35L, 36L, 37L, 38L, 39L, 40L, 40L, 40L, 43L,         43L, 45L, 45L, 47L, 0L, 49L, 49L, 51L, 52L, 51L, 54L,         36L, 0L, 57L, 58L), visible = c(TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, FALSE, FALSE,         FALSE), namespace = c("pkgdown", "pkgdown", "pkgdown",         "pkgdown", "pkgdown", "base", "purrr", "purrr", "purrr",         "base", "purrr", "pkgdown", "base", "pkgdown", "pkgdown",         "pkgdown", "downlit", "evaluate", "base", "base", "base",         "base", "base", "base", "base", "base", "base", "base",         "evaluate", "base", "base", "base", "base", "base", "base",         "tyler", "logger", "logger", "base", "logger", "base",         "base", "logger", "base", "logger", "base", "glue", "glue",         "glue", NA, "glue", "glue", "base", "base", "base", "base",         "base", "glue", "rlang"), scope = c("::", "::", ":::",         "::", ":::", "::", "::", ":::", ":::", "::", ":::", "local",         "::", ":::", ":::", ":::", "::", "::", "::", "local",         "local", "local", "local", "local", "local", "local",         "local", "local", ":::", "::", "::", "::", "::", "::",         "::", "::", "::", "::", "::", "local", "::", "::", "local",         "::", "local", "::", "::", "::", "local", NA, "local",         ":::", "::", "::", "::", "::", "::", "local", "::")), row.names = c(NA,     -59L), version = 2L, class = c("rlang_trace", "rlib_trace",     "tbl", "data.frame")), parent = structure(list(message = "object 'taxonomy_and_aaos_tbl' not found",         call = eval(expr, envir)), class = c("simpleError", "error",     "condition")), rlang = list(inherit = TRUE), call = NULL), class = c("rlang_error", "error", "condition"))): `glue` failed in `formatter_glue` on:
#> 
#>  chr "Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns."
#> 
#> Raw error message:
#> 
#> Failed to evaluate glue component {nrow(taxonomy_and_aaos_tbl)}
#> Caused by error:
#> ! object 'taxonomy_and_aaos_tbl' not found
#> 
#> Please consider using another `log_formatter` or `skip_formatter` on strings with curly braces.

# Example 2: Custom output path
sampled_data <- phase0_create_sampled_by_region(taxonomy_and_aaos_tbl, census_bureau_tbl,
                                                output_csv_path = "custom_path/sample.csv")
#> INFO [2024-11-18 00:43:04] Starting phase0_create_sampled_by_region.
#> Error in (function (e) {    args <- paste0(capture.output(str(...)), collapse = "\n")    stop(paste0("`glue` failed in `formatter_glue` on:\n\n",         args, "\n\nRaw error message:\n\n", conditionMessage(e),         "\n\nPlease consider using another `log_formatter` or ",         "`skip_formatter` on strings with curly braces."))})(structure(list(message = "Failed to evaluate glue component {nrow(taxonomy_and_aaos_tbl)}",     trace = structure(list(call = list(pkgdown::build_site_github_pages(new_process = FALSE,         install = FALSE), build_site(pkg, preview = FALSE, install = install,         new_process = new_process, ...), build_site_local(pkg = pkg,         examples = examples, run_dont_run = run_dont_run, seed = seed,         lazy = lazy, override = override, preview = preview,         devel = devel), build_reference(pkg, lazy = lazy, examples = examples,         run_dont_run = run_dont_run, seed = seed, override = override,         preview = FALSE, devel = devel), unwrap_purrr_error(purrr::map(topics,         build_reference_topic, pkg = pkg, lazy = lazy, examples_env = examples_env,         run_dont_run = run_dont_run)), withCallingHandlers(code,         purrr_error_indexed = function(err) {            cnd_signal(err$parent)        }), purrr::map(topics, build_reference_topic, pkg = pkg,         lazy = lazy, examples_env = examples_env, run_dont_run = run_dont_run),         map_("list", .x, .f, ..., .progress = .progress), with_indexed_errors(i = i,             names = names, error_call = .purrr_error_call, call_with_cleanup(map_impl,                 environment(), .type, .progress, n, names, i)),         withCallingHandlers(expr, error = function(cnd) {            if (i == 0L) {            }            else {                message <- c(i = "In index: {i}.")                if (!is.null(names) && !is.na(names[[i]]) &&                   names[[i]] != "") {                  name <- names[[i]]                  message <- c(message, i = "With name: {name}.")                }                else {                  name <- NULL                }                cli::cli_abort(message, location = i, name = name,                   parent = cnd, call = error_call, class = "purrr_error_indexed")            }        }), call_with_cleanup(map_impl, environment(), .type,             .progress, n, names, i), .f(.x[[i]], ...), withCallingHandlers(data_reference_topic(topic,             pkg, examples_env = examples_env, run_dont_run = run_dont_run),             error = function(err) {                cli::cli_abort("Failed to parse Rd in {.file {topic$file_in}}",                   parent = err, call = quote(build_reference()))            }), data_reference_topic(topic, pkg, examples_env = examples_env,             run_dont_run = run_dont_run), run_examples(tags$tag_examples[[1]],             env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),             topic = tools::file_path_sans_ext(topic$file_in),             run_dont_run = run_dont_run), highlight_examples(code,             topic, env = env), downlit::evaluate_and_highlight(code,             fig_save = fig_save_topic, env = eval_env, output_handler = handler),         evaluate::evaluate(code, child_env(env), new_device = TRUE,             output_handler = output_handler), withRestarts(with_handlers({            for (expr in tle$exprs) {                ev <- withVisible(eval(expr, envir))                watcher$capture_plot_and_output()                watcher$print_value(ev$value, ev$visible, envir)            }            TRUE        }, handlers), eval_continue = function() TRUE, eval_stop = function() FALSE,             eval_error = function(cnd) {                signalCondition(cnd)                stop(cnd)            }), withRestartList(expr, restarts), withOneRestart(withRestartList(expr,             restarts[-nr]), restarts[[nr]]), doWithOneRestart(return(expr),             restart), withRestartList(expr, restarts[-nr]), withOneRestart(withRestartList(expr,             restarts[-nr]), restarts[[nr]]), doWithOneRestart(return(expr),             restart), withRestartList(expr, restarts[-nr]), withOneRestart(expr,             restarts[[1L]]), doWithOneRestart(return(expr), restart),         with_handlers({            for (expr in tle$exprs) {                ev <- withVisible(eval(expr, envir))                watcher$capture_plot_and_output()                watcher$print_value(ev$value, ev$visible, envir)            }            TRUE        }, handlers), eval(call), eval(call), withCallingHandlers(code,             message = `<fn>`, warning = `<fn>`, error = `<fn>`),         withVisible(eval(expr, envir)), eval(expr, envir), eval(expr,             envir), phase0_create_sampled_by_region(taxonomy_and_aaos_tbl,             census_bureau_tbl, output_csv_path = "custom_path/sample.csv"),         logger::log_info("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns."),         log_level(INFO, ..., namespace = namespace, .logcall = .logcall,             .topcall = .topcall, .topenv = .topenv), lapply(definitions,             function(definition) {                if (level > definition$threshold) {                  return(NULL)                }                log_fun <- do.call(logger, definition)                structure(do.call(log_fun, log_arg), class = "logger")            }), FUN(X[[i]], ...), structure(do.call(log_fun,             log_arg), class = "logger"), do.call(log_fun, log_arg),         `<fn>`("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns.",             level = `<loglevel>`, .logcall = logger::log_info("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns."),             .topcall = phase0_create_sampled_by_region(taxonomy_and_aaos_tbl,                 census_bureau_tbl, output_csv_path = "custom_path/sample.csv"),             .topenv = `<env>`, namespace = "tyler"), do.call(formatter,             c(res$params, list(.logcall = substitute(.logcall),                 .topcall = substitute(.topcall), .topenv = .topenv))),         `<fn>`("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns.",             .logcall = logger::log_info("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns."),             .topcall = phase0_create_sampled_by_region(taxonomy_and_aaos_tbl,                 census_bureau_tbl, output_csv_path = "custom_path/sample.csv"),             .topenv = `<env>`), withCallingHandlers(glue::glue(...,             .envir = .topenv), error = function(e) {            args <- paste0(capture.output(str(...)), collapse = "\n")            stop(paste0("`glue` failed in `formatter_glue` on:\n\n",                 args, "\n\nRaw error message:\n\n", conditionMessage(e),                 "\n\nPlease consider using another `log_formatter` or ",                 "`skip_formatter` on strings with curly braces."))        }), glue::glue(..., .envir = .topenv), glue_data(.x = NULL,             ..., .sep = .sep, .envir = .envir, .open = .open,             .close = .close, .na = .na, .null = .null, .comment = .comment,             .literal = .literal, .transformer = .transformer,             .trim = .trim), `<fn>`("nrow(taxonomy_and_aaos_tbl)"),         .transformer(expr, env) %||% .null, .transformer(expr,             env), with_glue_error(eval(expr, envir), paste0("Failed to evaluate glue component {",             text, "}")), withCallingHandlers(expr, error = function(cnd) {            rlang::abort(message, parent = cnd, call = NULL)        }), eval(expr, envir), eval(expr, envir), nrow(taxonomy_and_aaos_tbl),         .handleSimpleError(`<fn>`, "object 'taxonomy_and_aaos_tbl' not found",             base::quote(eval(expr, envir))), h(simpleError(msg,             call)), rlang::abort(message, parent = cnd, call = NULL)),         parent = c(0L, 1L, 2L, 3L, 4L, 5L, 4L, 7L, 8L, 9L, 8L,         8L, 12L, 12L, 14L, 15L, 16L, 17L, 18L, 19L, 20L, 21L,         20L, 23L, 24L, 23L, 26L, 27L, 18L, 29L, 30L, 29L, 18L,         18L, 34L, 35L, 36L, 37L, 38L, 39L, 40L, 40L, 40L, 43L,         43L, 45L, 45L, 47L, 0L, 49L, 49L, 51L, 52L, 51L, 54L,         36L, 0L, 57L, 58L), visible = c(TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, FALSE, FALSE,         FALSE), namespace = c("pkgdown", "pkgdown", "pkgdown",         "pkgdown", "pkgdown", "base", "purrr", "purrr", "purrr",         "base", "purrr", "pkgdown", "base", "pkgdown", "pkgdown",         "pkgdown", "downlit", "evaluate", "base", "base", "base",         "base", "base", "base", "base", "base", "base", "base",         "evaluate", "base", "base", "base", "base", "base", "base",         "tyler", "logger", "logger", "base", "logger", "base",         "base", "logger", "base", "logger", "base", "glue", "glue",         "glue", NA, "glue", "glue", "base", "base", "base", "base",         "base", "glue", "rlang"), scope = c("::", "::", ":::",         "::", ":::", "::", "::", ":::", ":::", "::", ":::", "local",         "::", ":::", ":::", ":::", "::", "::", "::", "local",         "local", "local", "local", "local", "local", "local",         "local", "local", ":::", "::", "::", "::", "::", "::",         "::", "::", "::", "::", "::", "local", "::", "::", "local",         "::", "local", "::", "::", "::", "local", NA, "local",         ":::", "::", "::", "::", "::", "::", "local", "::")), row.names = c(NA,     -59L), version = 2L, class = c("rlang_trace", "rlib_trace",     "tbl", "data.frame")), parent = structure(list(message = "object 'taxonomy_and_aaos_tbl' not found",         call = eval(expr, envir)), class = c("simpleError", "error",     "condition")), rlang = list(inherit = TRUE), call = NULL), class = c("rlang_error", "error", "condition"))): `glue` failed in `formatter_glue` on:
#> 
#>  chr "Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns."
#> 
#> Raw error message:
#> 
#> Failed to evaluate glue component {nrow(taxonomy_and_aaos_tbl)}
#> Caused by error:
#> ! object 'taxonomy_and_aaos_tbl' not found
#> 
#> Please consider using another `log_formatter` or `skip_formatter` on strings with curly braces.

# Example 3: Using a subset of census_bureau_tbl for faster testing
sampled_data <- phase0_create_sampled_by_region(taxonomy_and_aaos_tbl,
                                                dplyr::filter(census_bureau_tbl, Region == "West"),
                                                output_csv_path = "test/sample.csv")
#> INFO [2024-11-18 00:43:04] Starting phase0_create_sampled_by_region.
#> Error in (function (e) {    args <- paste0(capture.output(str(...)), collapse = "\n")    stop(paste0("`glue` failed in `formatter_glue` on:\n\n",         args, "\n\nRaw error message:\n\n", conditionMessage(e),         "\n\nPlease consider using another `log_formatter` or ",         "`skip_formatter` on strings with curly braces."))})(structure(list(message = "Failed to evaluate glue component {nrow(taxonomy_and_aaos_tbl)}",     trace = structure(list(call = list(pkgdown::build_site_github_pages(new_process = FALSE,         install = FALSE), build_site(pkg, preview = FALSE, install = install,         new_process = new_process, ...), build_site_local(pkg = pkg,         examples = examples, run_dont_run = run_dont_run, seed = seed,         lazy = lazy, override = override, preview = preview,         devel = devel), build_reference(pkg, lazy = lazy, examples = examples,         run_dont_run = run_dont_run, seed = seed, override = override,         preview = FALSE, devel = devel), unwrap_purrr_error(purrr::map(topics,         build_reference_topic, pkg = pkg, lazy = lazy, examples_env = examples_env,         run_dont_run = run_dont_run)), withCallingHandlers(code,         purrr_error_indexed = function(err) {            cnd_signal(err$parent)        }), purrr::map(topics, build_reference_topic, pkg = pkg,         lazy = lazy, examples_env = examples_env, run_dont_run = run_dont_run),         map_("list", .x, .f, ..., .progress = .progress), with_indexed_errors(i = i,             names = names, error_call = .purrr_error_call, call_with_cleanup(map_impl,                 environment(), .type, .progress, n, names, i)),         withCallingHandlers(expr, error = function(cnd) {            if (i == 0L) {            }            else {                message <- c(i = "In index: {i}.")                if (!is.null(names) && !is.na(names[[i]]) &&                   names[[i]] != "") {                  name <- names[[i]]                  message <- c(message, i = "With name: {name}.")                }                else {                  name <- NULL                }                cli::cli_abort(message, location = i, name = name,                   parent = cnd, call = error_call, class = "purrr_error_indexed")            }        }), call_with_cleanup(map_impl, environment(), .type,             .progress, n, names, i), .f(.x[[i]], ...), withCallingHandlers(data_reference_topic(topic,             pkg, examples_env = examples_env, run_dont_run = run_dont_run),             error = function(err) {                cli::cli_abort("Failed to parse Rd in {.file {topic$file_in}}",                   parent = err, call = quote(build_reference()))            }), data_reference_topic(topic, pkg, examples_env = examples_env,             run_dont_run = run_dont_run), run_examples(tags$tag_examples[[1]],             env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),             topic = tools::file_path_sans_ext(topic$file_in),             run_dont_run = run_dont_run), highlight_examples(code,             topic, env = env), downlit::evaluate_and_highlight(code,             fig_save = fig_save_topic, env = eval_env, output_handler = handler),         evaluate::evaluate(code, child_env(env), new_device = TRUE,             output_handler = output_handler), withRestarts(with_handlers({            for (expr in tle$exprs) {                ev <- withVisible(eval(expr, envir))                watcher$capture_plot_and_output()                watcher$print_value(ev$value, ev$visible, envir)            }            TRUE        }, handlers), eval_continue = function() TRUE, eval_stop = function() FALSE,             eval_error = function(cnd) {                signalCondition(cnd)                stop(cnd)            }), withRestartList(expr, restarts), withOneRestart(withRestartList(expr,             restarts[-nr]), restarts[[nr]]), doWithOneRestart(return(expr),             restart), withRestartList(expr, restarts[-nr]), withOneRestart(withRestartList(expr,             restarts[-nr]), restarts[[nr]]), doWithOneRestart(return(expr),             restart), withRestartList(expr, restarts[-nr]), withOneRestart(expr,             restarts[[1L]]), doWithOneRestart(return(expr), restart),         with_handlers({            for (expr in tle$exprs) {                ev <- withVisible(eval(expr, envir))                watcher$capture_plot_and_output()                watcher$print_value(ev$value, ev$visible, envir)            }            TRUE        }, handlers), eval(call), eval(call), withCallingHandlers(code,             message = `<fn>`, warning = `<fn>`, error = `<fn>`),         withVisible(eval(expr, envir)), eval(expr, envir), eval(expr,             envir), phase0_create_sampled_by_region(taxonomy_and_aaos_tbl,             dplyr::filter(census_bureau_tbl, Region == "West"),             output_csv_path = "test/sample.csv"), logger::log_info("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns."),         log_level(INFO, ..., namespace = namespace, .logcall = .logcall,             .topcall = .topcall, .topenv = .topenv), lapply(definitions,             function(definition) {                if (level > definition$threshold) {                  return(NULL)                }                log_fun <- do.call(logger, definition)                structure(do.call(log_fun, log_arg), class = "logger")            }), FUN(X[[i]], ...), structure(do.call(log_fun,             log_arg), class = "logger"), do.call(log_fun, log_arg),         `<fn>`("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns.",             level = `<loglevel>`, .logcall = logger::log_info("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns."),             .topcall = phase0_create_sampled_by_region(taxonomy_and_aaos_tbl,                 dplyr::filter(census_bureau_tbl, Region == "West"),                 output_csv_path = "test/sample.csv"), .topenv = `<env>`,             namespace = "tyler"), do.call(formatter, c(res$params,             list(.logcall = substitute(.logcall), .topcall = substitute(.topcall),                 .topenv = .topenv))), `<fn>`("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns.",             .logcall = logger::log_info("Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns."),             .topcall = phase0_create_sampled_by_region(taxonomy_and_aaos_tbl,                 dplyr::filter(census_bureau_tbl, Region == "West"),                 output_csv_path = "test/sample.csv"), .topenv = `<env>`),         withCallingHandlers(glue::glue(..., .envir = .topenv),             error = function(e) {                args <- paste0(capture.output(str(...)), collapse = "\n")                stop(paste0("`glue` failed in `formatter_glue` on:\n\n",                   args, "\n\nRaw error message:\n\n", conditionMessage(e),                   "\n\nPlease consider using another `log_formatter` or ",                   "`skip_formatter` on strings with curly braces."))            }), glue::glue(..., .envir = .topenv), glue_data(.x = NULL,             ..., .sep = .sep, .envir = .envir, .open = .open,             .close = .close, .na = .na, .null = .null, .comment = .comment,             .literal = .literal, .transformer = .transformer,             .trim = .trim), `<fn>`("nrow(taxonomy_and_aaos_tbl)"),         .transformer(expr, env) %||% .null, .transformer(expr,             env), with_glue_error(eval(expr, envir), paste0("Failed to evaluate glue component {",             text, "}")), withCallingHandlers(expr, error = function(cnd) {            rlang::abort(message, parent = cnd, call = NULL)        }), eval(expr, envir), eval(expr, envir), nrow(taxonomy_and_aaos_tbl),         .handleSimpleError(`<fn>`, "object 'taxonomy_and_aaos_tbl' not found",             base::quote(eval(expr, envir))), h(simpleError(msg,             call)), rlang::abort(message, parent = cnd, call = NULL)),         parent = c(0L, 1L, 2L, 3L, 4L, 5L, 4L, 7L, 8L, 9L, 8L,         8L, 12L, 12L, 14L, 15L, 16L, 17L, 18L, 19L, 20L, 21L,         20L, 23L, 24L, 23L, 26L, 27L, 18L, 29L, 30L, 29L, 18L,         18L, 34L, 35L, 36L, 37L, 38L, 39L, 40L, 40L, 40L, 43L,         43L, 45L, 45L, 47L, 0L, 49L, 49L, 51L, 52L, 51L, 54L,         36L, 0L, 57L, 58L), visible = c(TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE,         TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, FALSE, FALSE,         FALSE), namespace = c("pkgdown", "pkgdown", "pkgdown",         "pkgdown", "pkgdown", "base", "purrr", "purrr", "purrr",         "base", "purrr", "pkgdown", "base", "pkgdown", "pkgdown",         "pkgdown", "downlit", "evaluate", "base", "base", "base",         "base", "base", "base", "base", "base", "base", "base",         "evaluate", "base", "base", "base", "base", "base", "base",         "tyler", "logger", "logger", "base", "logger", "base",         "base", "logger", "base", "logger", "base", "glue", "glue",         "glue", NA, "glue", "glue", "base", "base", "base", "base",         "base", "glue", "rlang"), scope = c("::", "::", ":::",         "::", ":::", "::", "::", ":::", ":::", "::", ":::", "local",         "::", ":::", ":::", ":::", "::", "::", "::", "local",         "local", "local", "local", "local", "local", "local",         "local", "local", ":::", "::", "::", "::", "::", "::",         "::", "::", "::", "::", "::", "local", "::", "::", "local",         "::", "local", "::", "::", "::", "local", NA, "local",         ":::", "::", "::", "::", "::", "::", "local", "::")), row.names = c(NA,     -59L), version = 2L, class = c("rlang_trace", "rlib_trace",     "tbl", "data.frame")), parent = structure(list(message = "object 'taxonomy_and_aaos_tbl' not found",         call = eval(expr, envir)), class = c("simpleError", "error",     "condition")), rlang = list(inherit = TRUE), call = NULL), class = c("rlang_error", "error", "condition"))): `glue` failed in `formatter_glue` on:
#> 
#>  chr "Inputs provided: taxonomy_and_aaos_tbl with {nrow(taxonomy_and_aaos_tbl)} rows and {ncol(taxonomy_and_aaos_tbl)} columns."
#> 
#> Raw error message:
#> 
#> Failed to evaluate glue component {nrow(taxonomy_and_aaos_tbl)}
#> Caused by error:
#> ! object 'taxonomy_and_aaos_tbl' not found
#> 
#> Please consider using another `log_formatter` or `skip_formatter` on strings with curly braces.