Retrieve Clinician Data

Retrieves clinician data for each valid NPI from a dataframe or CSV file. This function ensures valid NPI numbers are processed and retrieves detailed clinician information.

Usage

retrieve_clinician_data(npi_numbers)

Arguments

input_data

Either a dataframe containing NPI numbers (must include a column named npi) or a path to a CSV file with NPI numbers.

Value

A tibble with clinician data for each valid NPI.

Details

The function checks the input data for validity and removes invalid NPI numbers.

If a CSV file path is provided, the file is read, and the data is processed.

Examples

# Example 1: Provide a dataframe with NPI numbers
sample_data <- tibble::tibble(npi = c(1689603763, 1234567890))
clinician_data <- retrieve_clinician_data(sample_data)
#> Error in purrr::map(npi_numbers, fetch_clinician_data): ℹ In index: 1.
#> ℹ With name: npi.
#> Caused by error in `h()`:
#> ! `glue` failed in `formatter_glue` on:
#> 
#>  'glue' chr [1:2] "Fetching data for NPI: 1689603763" ...
#> 
#> Raw error message:
#> 
#> All unnamed arguments must be length 1
#> 
#> Please consider using another `log_formatter` or `skip_formatter` on strings with curly braces.
print(clinician_data)
#> Error: object 'clinician_data' not found

# Example 2: Provide a CSV file path
write.csv(sample_data, "npi_data.csv", row.names = FALSE)
clinician_data <- retrieve_clinician_data("npi_data.csv")
#> INFO [2024-12-14 20:38:43] Fetching data for NPI: npi_data.csv
#> 10 records requested
#> Requesting records 0-10...
#> ERROR [2024-12-14 20:38:44] Error retrieving data for NPI npi_data.csv: 
#> Field: number
#> NPI must be 10 digits
#> INFO [2024-12-14 20:38:44] Data retrieved for 0 NPIs
print(clinician_data)
#> list()