Count Physicians by State or Subdivision

This function counts the number of physicians available per state or US Census Bureau subdivision. It can optionally save the counts to a CSV file.

Usage

count_physicians_by_group(
  data,
  state_name_column = "state_code",
  phone_column = "phone_number",
  first_name_column = "first",
  last_name_column = "last",
  group_by = "state",
  output_to_csv = NULL
)

Arguments

data

A dataframe containing the physician data.

state_name_column

The column name in the dataframe containing state names or abbreviations (default is "state_code").

phone_column

The column name in the dataframe containing phone numbers (default is "phone_number").

first_name_column

The column name in the dataframe containing first names (default is "first").

last_name_column

The column name in the dataframe containing last names (default is "last").

group_by

A string indicating whether to group by "state" or "subdivision" (default is "state").

output_to_csv

(Optional) A file path to save the state/subdivision counts to a CSV file. If NULL, no file is saved.

Value

A tibble with the counts of physicians per state or subdivision.

Examples

# Example 1: Count physicians by state
count_physicians_by_group(taxonomy_and_aaos_data)
#> Error: Failed to evaluate glue component {nrow(data)}
#> Caused by error:
#> ! object 'taxonomy_and_aaos_data' not found

# Example 2: Count physicians by U.S. Census Bureau subdivision
count_physicians_by_group(taxonomy_and_aaos_data, group_by = "subdivision")
#> Error: Failed to evaluate glue component {nrow(data)}
#> Caused by error:
#> ! object 'taxonomy_and_aaos_data' not found