Search NPI by Taxonomy

This function performs a search on the NPI registry for a list of taxonomy descriptions and returns a combined dataframe of results. It processes the returned data and cleans up unnecessary fields.

search_by_taxonomy(taxonomy_to_search)

Arguments

taxonomy_to_search

A character vector of taxonomy descriptions to search for in the NPI registry.

Value

A dataframe of cleaned NPI data, filtered by taxonomy and other criteria.