Split data into multiple parts and save each part as separate CSV files

This function splits the data based on a specified column (e.g., insurance or specialty), and saves each part as a separate CSV file. It allows you to prioritize certain values in the column, and ensures that no physician appears in consecutive rows.

split_calls_to_lab_assistants_and_save_by_priority(
  input_data_or_path,
  output_folder = getwd(),
  assistant_names,
  seed_value = 1978,
  complete_file_prefix = "complete_version_",
  split_file_prefix = "",
  recursive_create = TRUE,
  split_column = "insurance",
  priority_values = NULL
)

Arguments

input_data_or_path

Either a dataframe containing the input data or a path to the input data file (RDS, CSV, or XLS/XLSX).

output_folder

Directory where output CSV files will be saved. Default is current working directory.

assistant_names

Vector of lab assistant names to name the output files.

seed_value

Seed value for randomization (default is 1978).

complete_file_prefix

Prefix for the complete output file name (default is "complete_version_").

split_file_prefix

Prefix for each split output file name (default is empty).

recursive_create

Logical indicating if directories should be created recursively (default is TRUE).

split_column

The column on which to base the split (e.g., "insurance", "specialty_primary"). Default is "insurance".

priority_values

Vector of values to prioritize within the split_column (e.g., c("Pediatric Dermatologists", "Medicaid")). Default is NULL.

Examples

if (FALSE) { # \dontrun{
split_calls_to_lab_assistants_and_save_by_priority(input_data_or_path = "/path/to/file.csv",
                                                  output_folder = "/path/to/output/",
                                                  assistant_names = c("Assistant1", "Assistant2"),
                                                  split_column = "specialty_primary",
                                                  priority_values = c("Pediatric Dermatologists"))
} # }